fix(btree): 修复并发数据丢失和 goroutine 泄漏问题
## 问题概述
在高并发场景下发现严重的数据丢失和资源泄漏问题：
- 并发插入丢失率 2.42% (242/10000 操作)
- Goroutine 泄漏导致线程耗尽
- 测试失败率高达 89%

## 核心修复

### 1. 并发插入数据丢失修复
**问题**: PageRefCache 的 Delete() + Update() 存在 TOCTOU 竞态条件

**解决方案**:
- 新增原子 `Replace()` 方法：在同一锁保护下完成 Delete + Update
- 区分 UpdateLeafEntry 和 split 场景：
  - UpdateLeafEntry 成功：直接返回 nil，避免不必要重试
  - Split 场景：返回 ErrRetry 重新搜索路径
- 修复 epoch 推进时机：只在真正成功时推进

**影响**: 并发成功率从 97.58% 提升到 100%

### 2. 错误包装破坏重试机制
**问题**: `fmt.Errorf("find leaf ref: %w", err)` 包装导致 errors.Is() 失败

**解决方案**: 移除所有错误包装，直接返回原始 err
- search_path.go:219
- leaf_lock_set.go:50
- btree.go:611, 901

**影响**: setDirect() 的 10 次重试机制恢复正常工作

### 3. Goroutine 泄漏修复
**问题**: scheduleDelayedRelease() 启动的 goroutines 没有生命周期管理

**解决方案**:
- BTree 结构添加 ctx 和 cancelFunc 字段
- RootPageRef 添加 context 支持
- Close() 时调用 cancelFunc 取消所有后台 goroutines
- scheduleDelayedRelease 使用 context 监听取消信号

**影响**: 100 次运行 × 5 轮测试全部通过，无资源泄漏

## 测试验证

### 单元测试
- TestSetWithLeafLock_Concurrent: 20/20 轮通过
- TestSetWithLeafLock_ExtremeConcurrency: 20/20 轮通过
- TestGoroutineLeakFix: 新增验证测试
- 318 个 B-tree 测试全部通过

### 集成测试
- 完整测试套件 (make test): 通过
- Race detector: 无竞态条件
- 并发压力测试: 通过

## 性能影响
- 并发成功率: 97.58% → 100% (+2.42%)
- 数据丢失: 242/10000 → 0/10000
- Goroutine 泄漏: 已修复
- 测试稳定性: 显著提升

## 修改文件
- btree.go: context 管理、注释优化
- btree_ops.go: epoch 推进注释
- leaf_lock_set.go: 场景区分、错误处理
- search_path.go: 移除错误包装
- root_page_ref.go: context 支持、goroutine 管理
- failure_log_collection_test.go: 优化测试参数
- goroutine_leak_test.go: 新增验证测试
- ccow_manager_test.go: 更新构造函数调用

## 相关 Issue
- 循环引用和数据丢失根本原因调查
- 并发场景下的数据完整性保障
- 资源管理和生命周期控制

+++ internal/infrastructure/storage/btree/offheap/page_init_test.go
@@ -43,7 +43,7 @@ func TestInitPage_ClearsExtraChild(t *testing.T) {
 
 	// 5. 验证 extraChild 被清空
 	newHeader := pa.GetHeader(newPageID)
-	assert.Equal(t, uint32(0), newHeader.extraChild,
+	assert.Equal(t, uint64(0), newHeader.extraChild,
 		"extraChild should be 0 after InitPage")
 }
 
@@ -63,6 +63,6 @@ func TestInitIndexPage_CreatesValidPage(t *testing.T) {
 	header := pa.GetHeader(pageID)
 	assert.Equal(t, uint8(PageTypeIndex), header.pageType)
 	assert.Equal(t, uint16(0), header.count)
-	assert.Equal(t, uint32(0), header.extraChild, "extraChild should be 0")
+	assert.Equal(t, uint64(0), header.extraChild, "extraChild should be 0")
 	assert.Equal(t, uint64(1), header.version)
 }
